requirements.txt
schedule==1.1.0
python-dotenv==0.20.0
pandas==1.4.3
matplotlib==3.4.3
entsoe-py==0.5.8
tweepy==4.10.1
Pillow==8.4.0
pymongo==4.2.0
dnspython==2.2.1